Stream data using Apache bucket brigades
As a legacy from Noze.io the ServerResponse object
already was setup as a stream. Up to now it was using
ap_rwrite etc to directly write to the Apache socket.

This change converts it to a proper Apache stream.
That is, a writev() creates a proper brigade and
passed that down to the Apache 2 filter chain.

TODO: Create an own bucket type for arrays.