2-in-1: Improve OPTIONS, add onMount event
The cors() middleware doesn't stop the stack
when encountering an OPTIONS request. Instead
it lets it flow through (can be set via arg).
In addition, the Express final handler doesn't
treat OPTIONS as an error but just delivers it.

Also: Add the concept of a 'mountable middleware
object'. Nothing special about it, it can receive
onMount events (and by that get access to the
parent application)