Add a bodyParser.davQuery for WebDAV requests
This is a middleware which can parse the request body
of PROPFIND and some REPORT requests. The results can
be accessed as request.davQuery.

DAV requests can look complex but are all actually
pretty similar. The have a depth, properties, sometimes
a sync-token or some extra URLs.

This could be done a little nicer, but it'll do for
now. It's not great but not aweful either :-)