📆 Add CalDAV support to mods_todomvc
This commit adds CalDAV support to our mods_todomvc, our
Todo-Backend demo.

The regular Todo-Backend is left alone in the TodoMVCMain.swift
file. All the CalDAV stuff is wrapped in its own middleware.
As with any demo the code is not as beautiful as it should be ;->

Also included is a raw Apache handler to handle .well-known
requests which make the bootstrapping of DAV accounts easier.

As a final bonus: A tinzy CardDAV client serving three embedded
cards.