Hide Apache imp, add type-safe SQL select ext
Only expose the SQL interfaces, not the actual
Apache DBD implementation. Due to Swift protocol
limitations it can't be as awesome as desired.

Also add an extension which allows typesafe
SQL fetches using generics along the lines of

    dbd.select("SELECT * FROM pets") {
      ( name : String, count : Int? ) in
      req.puts("<li>\(name) (\(count))</li>")
    }

and

    struct Model {
      struct Pet {
        static let name  = Attribute<String>(name: "name")
        static let count = Attribute<Int>   (name: "count")
      }
    }
    dbd.select(Model.Pet.name, Model.Pet.count, from: "pets") {
      name, count in // types are derived from Attribute
      req.puts("<tr><td>\(name)</td><td>\(count)</td></tr>")
    }

The idea here is that SQL SELECTs have a fixed
schema and 'we' want the same in Swift. So it
would be cool to bring them together.
Traditional 'ORM's almost always rely on heavy
reflection and dynamic property access. The
stuff here is much more limited, but completely
typesafe.